+++ server/src/main.rs
@@ -4,7 +4,7 @@ mod net;
 use std::sync::{Arc, RwLock};
 use std::thread;
 
-const TICKRATE : u64 = 15; // ms
+const TICKRATE : u32 = 15; // ms
 
 pub mod schema_capnp {
     include!(concat!(env!("OUT_DIR"), "/schema_capnp.rs"));
@@ -18,8 +18,8 @@ fn main() {
 
     loop {
         let start = std::time::Instant::now();
-        gamestate.write().unwrap().tick(TICKRATE as f32);
-        thread::sleep(std::time::Duration::from_millis(TICKRATE));
+        gamestate.write().unwrap().tick(TICKRATE);
+        thread::sleep(std::time::Duration::from_millis(TICKRATE as u64));
         println!("tick [{}ms] - {} players", start.elapsed().as_millis(),  &gamestate.read().unwrap().players.len());
     }
 }
